+++ app/src/main/java/com/pvp/app/ui/screen/calendar/TaskFormsCommon.kt
@@ -1,41 +1,24 @@
 package com.pvp.app.ui.screen.calendar
 
-import androidx.compose.foundation.BasicTooltipBox
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.rememberBasicTooltipState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Info
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExposedDropdownMenuBox
 import androidx.compose.material3.ExposedDropdownMenuDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Slider
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
-import androidx.compose.material3.TooltipDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableDoubleStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
@@ -47,10 +30,10 @@ import com.pvp.app.ui.common.EditableInfoItem
 import com.pvp.app.ui.common.LabelFieldWrapper
 import com.pvp.app.ui.common.PickerPair
 import com.pvp.app.ui.common.PickerState
-import kotlinx.coroutines.launch
+import com.pvp.app.ui.common.InfoTooltip
 import java.time.Duration
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TaskEditFieldsSport(
     task: SportTask? = null,
@@ -110,45 +93,10 @@ fun TaskEditFieldsSport(
                 Modifier.fillMaxWidth(),
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                val tooltipState = rememberBasicTooltipState()
-                val scope = rememberCoroutineScope()
-
                 Text(activity?.title ?: "")
 
                 if (activity?.supportsDistanceMetrics == true) {
                     InfoTooltip(tooltipText = "This task is likely to be autocompleted")
-                    /*BasicTooltipBox(
-                        positionProvider = TooltipDefaults.rememberRichTooltipPositionProvider(),
-                        tooltip = {
-                            Text(
-                                text = "This task is likely to be autocompleted",
-                                modifier = Modifier
-                                    .clip(MaterialTheme.shapes.medium)
-                                    .background(MaterialTheme.colorScheme.primary)
-                                    .border(
-                                        border = BorderStroke(
-                                            1.dp,
-                                            MaterialTheme.colorScheme.outline
-                                        ),
-                                        shape = MaterialTheme.shapes.medium,
-                                    )
-                                    .height(40.dp)
-                                    .padding(8.dp)
-                                    .wrapContentSize(Alignment.Center),
-                                color = Color.White
-                            )
-                        },
-                        state = tooltipState
-                    ) {
-                        IconButton(
-                            onClick = { scope.launch { tooltipState.show() } },
-                        ) {
-                            Icon(
-                                imageVector = Icons.Outlined.Info,
-                                contentDescription = "Tooltip about autocompletion of activity"
-                            )
-                        }
-                    }*/
                 }
             }
         }
@@ -239,46 +187,4 @@ fun TaskEditFieldsSport(
     onDistanceChange(distance)
 
     duration?.let { onDurationChange(it) }
-}
-
-@OptIn(ExperimentalFoundationApi::class, ExperimentalMaterial3Api::class)
-@Composable
-fun InfoTooltip(
-    tooltipText: String
-) {
-    val tooltipState = rememberBasicTooltipState()
-    val scope = rememberCoroutineScope()
-
-    BasicTooltipBox(
-        positionProvider = TooltipDefaults.rememberRichTooltipPositionProvider(),
-        tooltip = {
-            Text(
-                text = tooltipText,
-                modifier = Modifier
-                    .clip(MaterialTheme.shapes.medium)
-                    .background(MaterialTheme.colorScheme.primary)
-                    .border(
-                        border = BorderStroke(
-                            1.dp,
-                            MaterialTheme.colorScheme.outline
-                        ),
-                        shape = MaterialTheme.shapes.medium,
-                    )
-                    .height(40.dp)
-                    .padding(8.dp)
-                    .wrapContentSize(Alignment.Center),
-                color = Color.White
-            )
-        },
-        state = tooltipState
-    ) {
-        IconButton(
-            onClick = { scope.launch { tooltipState.show() } },
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Info,
-                contentDescription = "Autocompletion of activity"
-            )
-        }
-    }
 }
